win/gem: address sizeof(int) review findings
- mar_nh_poskey: tail-recursion replaced with do-while; bounds stack
  growth if gcc doesn't TCO the FAIL path
- menuitem_invert_test: declared with signed char to match NetHack
  core's schar-typed boolean instead of E_GEM's enum-int boolean
- Max() narrowing: explicit short casts in mar_set_accelerators and
  mar_display_nhwindow NHW_MENU
- Gem_add_menu: explicit short casts for Gmi_attr and Gmi_color
- mar_set_accelerators: only overwrite Gmi_str[0] when it is the '?'
  placeholder Gem_add_menu writes

+++ win/gem/wingem.c
@@ -769,8 +769,8 @@ Gem_add_menu(winid window, const glyph_info *glyphinfo,
     G_item->Gmi_selected = (itemflags & MENU_ITEMFLAGS_SELECTED) ? 1 : 0;
     G_item->Gmi_accelerator = ch;
     G_item->Gmi_groupacc = gch;
-    G_item->Gmi_attr = attr;
-    G_item->Gmi_color = clr;
+    G_item->Gmi_attr = (short) attr;
+    G_item->Gmi_color = (short) clr;
     G_item->Gmi_itemflags =
         (unsigned short) (itemflags & ~MENU_ITEMFLAGS_SELECTED);
     G_item->Gmi_str = copy_of(newstr);
